pyproject.toml
[project]
name = "news-latter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.4.0",
    "feedparser>=6.0.12",
    "google-generativeai>=0.8.6",
    "lxml-html-clean>=0.4.4",
    "newspaper4k>=0.9.5",
    "python-dotenv>=1.2.2",
]
